% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodeCAGIS.R
\name{geocodeCAGIS}
\alias{geocodeCAGIS}
\title{Geocode Cincinnati, OH area address using an offline and parcel-based method}
\usage{
geocodeCAGIS(addr_string, return.score = TRUE, return.call = FALSE,
  return.match = TRUE)
}
\arguments{
\item{return.score}{logical, return method and matching score?}

\item{return.call}{logical, return the original address string?}

\item{return.match}{logical, return the best address text match from CAGIS?}

\item{address_string}{a single string that will be geocoded}
}
\value{
data.frame with lat/lon coords, CAGIS parcel id, and optionally method score, original call, or matched CAGIS record
}
\description{
Geocode an address using offline shapefile from CAGIS.
}
\details{
This function parses a given address string into address components and
attempts to match the address to CAGIS data.  The best match is returned and
the score represents how many insertions/deletions/rearrangements were needed
to match the input string to the address data.

The \code{sysdata.rda} file comes bundled with the package. Alternatively,
build the system data file on your own, using updated CAGIS files. See
the vignette for details on this operation.

This function will return NA if the zip code of the address string
does not begin with 450, 451, or 452.

Requires a sufficient python binary and the usaddress module.
See \link{addr_parse} for more details.
}
\examples{
# geocodeCAGIS('3333 Burnet Ave, Cincinnati, OH 45229')
# geocodeCAGIS('1456 Main St. 23566')
# geocodeCAGIS('3131 Mary Jane Dr 45211')
}
