% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_addresses.R
\name{parse_address}
\alias{parse_address}
\title{parse an address into components}
\usage{
parse_address(address)
}
\arguments{
\item{X}{an address as a character string}
}
\value{
data.frame of address parsing results
}
\description{
Note that this relies on python usaddress library. See the README for more details
}
\examples{
parse_address('3333 Burnet Ave, Cincinnati, OH 45229')
}
