% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_addresses.R
\name{parse_addresses}
\alias{parse_addresses}
\title{parse addresses into components}
\usage{
parse_addresses(X, ...)
}
\arguments{
\item{X}{a list or vector of addresses}

\item{...}{additional arguments passed to \code{CB::mappp}; set options for
cache and parallel operation here}
}
\value{
list of address parsing results with one address per element
}
\description{
Note that this relies on python with the ...
}
\examples{
parse_addresses(c('3333 Burnet Ave, Cincinnati, OH 45229', '737 US 50 Cincinnati OH 45150'))
}
